feat(@angular/ssr): improve handling of aborted requests in `AngularServerApp`
Introduce support for handling request signal abortions in the `AngularServerApp`. This is particularly useful in the development server integration where a 30-second timeout is enforced for requests/responses.
